fixed csv writer
It is now possible to append to existing file

+++ trading_methodologies/trading_util.py
@@ -2,11 +2,19 @@ import os
 import csv
 
 
-def write_as_csv(data): 
-    filename = 'trading_methodologies1.csv'  
-    #file_path = os.path.join('crawled_data', self.output_name, '.csv')
-    with open(filename, 'w+', newline='') as file:
-        writer = csv.writer(file)
-        writer.writerow(["Trading Method.", "Asset Alloc.", "Asset", "Amount($)", "Asset price", "# shares"])
-        for x in data:
-            writer.writerow(x)
+def write_as_csv(data, write_mode): 
+    filename = 'trading_methodologies.csv'  
+    #pass overwrite as argument to write the contents fresh
+    if write_mode == "overwrite":
+        with open(filename, 'w+', newline='') as file:
+            writer = csv.writer(file)
+            writer.writerow(["Trading Method.", "Asset Alloc.", "Asset", "Amount($)", "Asset price", "# shares"])
+            for x in data:
+                writer.writerow(x)
+    #or pass append if you just want to write to the end of the file
+    elif write_mode == "append":
+        with open(filename, 'a+', newline='') as file:
+            writer = csv.writer(file)
+            writer.writerow(["Trading Method.", "Asset Alloc.", "Asset", "Amount($)", "Asset price", "# shares"])
+            for x in data:
+                writer.writerow(x)
